requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2022.12.7
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
humanfriendly==10.0
idna==3.7
imageio==2.34.2
importlib-metadata==8.2.0
importlib-resources==6.4.0
jinja2==3.1.4
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy-loader==0.4
lightning==2.3.3
lightning-utilities==0.11.6
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
olefile==0.46
onnxruntime==1.18.1
opencv-python-headless==4.10.0.84
packaging==24.1
Pillow==9.0.1
pkgutil-resolve-name==1.3.10
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.2
psutil==6.0.0
PyMatting==1.1.12
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.3
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.35.1
rembg==2.0.58
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.21.0
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1635175459052/work
scikit-video==1.1.11
scipy==1.10.1
sentry-sdk==2.11.0
setproctitle==1.3.3
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.1
sympy==1.13.1
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tifffile==2023.7.10
torch==2.4.0
torchaudio==0.8.0a0+e4e171a
torchmetrics==1.4.0.post0
torchvision==0.9.1
tqdm @ file:///tmp/build/80754af9/tqdm_1627710282869/work
triton==3.0.0
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
urllib3==2.2.2
wandb==0.17.5
yarl==1.9.4
zipp==3.19.2
